Update RegisterSystem.py
RegisterSystem can now detect failed courses which are prerequisite to other courses and will add them into the list instead of the super-course. If there are any failed courses which are not prerequisite to any other course, register system will add them into the list as well.

It also adds random elective course(s) to the list if the student is in an available position.